Updated response code + listIncome uses limit + added updateIncome

+++ src/main/java/com/majicode/budgetapp/api/IncomesApiController.java
@@ -13,6 +13,8 @@ import javax.validation.constraints.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -35,13 +37,12 @@ import io.swagger.annotations.ApiParam;
 public class IncomesApiController implements IncomesApi {
 	private static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 	
+	private final IncomeRepository repository;
     private final NativeWebRequest request;
-    
-    @Autowired
-    private IncomeRepository repository;
 
-    @org.springframework.beans.factory.annotation.Autowired
-    public IncomesApiController(NativeWebRequest request) {
+    @Autowired
+    public IncomesApiController(IncomeRepository repository, NativeWebRequest request) {
+    	this.repository = repository;
         this.request = request;
     }
 
@@ -50,9 +51,19 @@ public class IncomesApiController implements IncomesApi {
         return Optional.ofNullable(request);
     }
     
+    /*
+      _____                             _               __  __      _   _               _     
+	 |  __ \                           (_)             |  \/  |    | | | |             | |    
+	 | |__) | __ ___   ___ ___  ___ ___ _ _ __   __ _  | \  / | ___| |_| |__   ___   __| |___ 
+	 |  ___/ '__/ _ \ / __/ _ \/ __/ __| | '_ \ / _` | | |\/| |/ _ \ __| '_ \ / _ \ / _` / __|
+	 | |   | | | (_) | (_|  __/\__ \__ \ | | | | (_| | | |  | |  __/ |_| | | | (_) | (_| \__ \
+	 |_|   |_|  \___/ \___\___||___/___/_|_| |_|\__, | |_|  |_|\___|\__|_| |_|\___/ \__,_|___/
+	                                             __/ |                                        
+	                                            |___/                                         
+     */
     @Override
     public ResponseEntity<Income> addIncome(@ApiParam(value = "Income definition" ,required=true )  @Valid @RequestBody Income income) {
-    	logger.info(">>> addIncome");
+    	logger.info("Adding a new Income");
     	
     	logger.debug("Income data:");
     	logger.debug("Name: {}", income.getName());
@@ -67,7 +78,7 @@ public class IncomesApiController implements IncomesApi {
     	savedIncome.setDateCreated(new java.sql.Date(new Date().getTime()));
     	savedIncome.setDateUpdated(savedIncome.getDateCreated());
     	
-    	logger.debug("Save the income");
+    	logger.info("Saving the income");
     	final IncomeData incomeData = repository.save(savedIncome);
     	
     	final UUID savedIncomeId = incomeData.getIncomeId();
@@ -83,44 +94,40 @@ public class IncomesApiController implements IncomesApi {
     @SuppressWarnings({ "rawtypes", "unchecked" })
 	@Override
     public ResponseEntity deleteIncome(@ApiParam(value = "Income id to delete",required=true) @PathVariable("incomeId") UUID incomeId) {
-    	logger.info(">>> deleteIncome");
+    	logger.info("Deleting an income");
     	
     	logger.debug("Parameters:");
     	logger.debug("incomeId: {}", incomeId);
     	
-    	logger.debug("Calling delete");
-    	final long initialCount = repository.count();
-    	repository.deleteById(incomeId);
-    	final boolean isDeleted = repository.count() == initialCount ? false : true;
+    	final Optional<IncomeData> income = repository.findById(incomeId);
     	
-    	logger.debug("isDeleted? {}", isDeleted);
-    	
-    	if (isDeleted) {
-    		return ResponseEntity.ok().build();
+    	if (!income.isPresent()) {
+    		final Error error = ApiUtil.generateError("Unable to delete the Income", 
+    				                                  "Income with the id " + incomeId + " not found");
+    		
+    		logError(error);
+    		
+    		return new ResponseEntity<Error>(error, HttpStatus.NOT_FOUND);
     	}
     	
-    	logger.error("ID was not found: {}", incomeId);
-		Error error = new Error();
-		error.setCode(HttpStatus.NOT_FOUND.value());
-		error.setMessage("Date was null");
-		
-		return new ResponseEntity<Error>(error, HttpStatus.NOT_FOUND);
+    	logger.debug("Calling delete");
+    	repository.deleteById(incomeId);
+    	return new ResponseEntity<Income>(HttpStatus.NO_CONTENT);
     }
 
     @SuppressWarnings({ "rawtypes", "unchecked" })
 	@Override
     public ResponseEntity findIncomeByDate(@NotNull @ApiParam(value = "Date the income was created", required = true) @Valid @RequestParam(value = "date", required = true) String date) {
-    	logger.info(">>> findIncomeByDate");
+    	logger.info("Find income by date");
     	
     	logger.debug("Parameters:");
     	logger.debug("Date: {}", date);
     	
     	logger.debug("Perform data NULL check");
     	if (date == null) {
-    		logger.error("Date was NULL");
-    		Error error = new Error();
-    		error.setCode(HttpStatus.BAD_REQUEST.value());
-    		error.setMessage("Date was null");
+    		final Error error = ApiUtil.generateError("Date is NULL", "");
+			
+			logError(error);
     		
     		return new ResponseEntity<Error>(error, HttpStatus.BAD_REQUEST);
     	}
@@ -158,18 +165,24 @@ public class IncomesApiController implements IncomesApi {
     
     @Override
     public ResponseEntity<List<Income>> listIncomes(@ApiParam(value = "How many Incomes to return at one time (max 100)", defaultValue = "0") @Valid @RequestParam(value = "limit", required = false, defaultValue="0") Optional<Integer> limit) {
-    	logger.info(">>> listIncomes");
+    	logger.info("List all incomes, with date created asc");
     	
     	logger.debug("Parameters:");
     	logger.debug("Limit: {}", limit.isPresent() ? limit.get() : "none provided");
     	
-    	// TODO: Need to implement
-//    	if (limit.isPresent()) {
-//    		final Pageable limitIncomes = PageRequest.of(0,limit.get());
-//    	}
+    	final List<IncomeData> fetchedIncomes = new ArrayList<>();
     	
-    	logger.debug("Call getIncomes");
-    	final List<IncomeData> fetchedIncomes = repository.findByOrderByDateCreatedAsc();
+    	logger.debug("Call findByOrderByDateCreatedAsc");
+    	if (limit.isPresent() && limit.get() != 0) {
+    		final Pageable limitIncomes = PageRequest.of(0, limit.get());
+    		fetchedIncomes.addAll(repository.findByOrderByDateCreatedAsc(limitIncomes));
+    	} else {
+    		fetchedIncomes.addAll(repository.findByOrderByDateCreatedAsc());
+    	}
+    	
+    	if (fetchedIncomes.isEmpty()) {
+    		return new ResponseEntity<List<Income>>(HttpStatus.NO_CONTENT);
+    	}
     	
     	logger.trace("Convert the incomes");
     	final List<Income> incomes = convert(fetchedIncomes);
@@ -180,20 +193,21 @@ public class IncomesApiController implements IncomesApi {
     @SuppressWarnings({ "rawtypes", "unchecked" })
 	@Override
     public ResponseEntity findIncomeById(@ApiParam(value = "The id of the income to retrieve",required=true) @PathVariable("incomeId") String incomeId) {
-    	logger.info(">>> findIncomeById");
+    	logger.info("Find income by id");
     	
     	logger.debug("Parameters:");
     	logger.debug("incomeId: {}", incomeId);
     	
     	logger.debug("Validate UUID");
+    	
     	// Check to see if input is in the proper UUID format
     	final boolean isUuid = ApiUtil.isValidUUID(incomeId);
     	
     	if (!isUuid) {
-    		logger.error("Not valid UUID");
-    		Error error = new Error();
-    		error.setCode(HttpStatus.BAD_REQUEST.value());
-    		error.setMessage("Income ID was not a valid UUID");
+    		final Error error = ApiUtil.generateError("Not a valid UUID provided", 
+    				                                  "Invalid UUID provided " + incomeId);
+    		
+    		logError(error);
     		
     		return new ResponseEntity<Error>(error, HttpStatus.BAD_REQUEST);
     	}
@@ -213,13 +227,57 @@ public class IncomesApiController implements IncomesApi {
     	
     	// Income was not found
     	logger.error("Income was not found with UUID: {}", incomeId);
-    	Error error = new Error();
-		error.setCode(HttpStatus.NOT_FOUND.value());
-		error.setMessage("Income was not found with id: " + incomeId);
+    	final Error error = ApiUtil.generateError("Income was not found with provided UUID", 
+                                                  "Invalid Income id provided " + incomeId);
+
+		logError(error);
 		
 		return new ResponseEntity<Error>(error, HttpStatus.BAD_REQUEST);
     }
     
+    @SuppressWarnings({ "rawtypes", "unchecked" })
+	public ResponseEntity updateIncome(@ApiParam(value = "Income object that needs to be updated in the budget" ,required=true )  @Valid @RequestBody Income income) {
+    	logger.info("Update an existing Income");
+    	
+    	final UUID uuid = income.getId();
+    	
+    	final Optional<IncomeData> incomeData = repository.findById(uuid);
+    	
+    	if (!incomeData.isPresent()) {
+    		logger.error("Income with id {} not found", income.getId());
+    		
+    		final Error error = ApiUtil.generateError("Income was not found with provided UUID", 
+                                                      "Invalid Income id provided " + uuid);
+
+    		logError(error);
+    		
+    		return new ResponseEntity<Error>(error, HttpStatus.NOT_FOUND);
+    	}
+    	
+    	final IncomeData currentIncome = incomeData.get();
+    	
+    	currentIncome.setDateUpdated(new java.sql.Date(new Date().getTime()));
+    	currentIncome.setName(income.getName());
+    	currentIncome.setPlannedAmount(income.getPlannedAmount());
+    	currentIncome.setReceivedAmount(income.getReceivedAmount());
+    	
+    	repository.save(currentIncome);
+    	
+    	final Income result = convert(currentIncome);
+    	
+    	return new ResponseEntity<Income>(result, HttpStatus.OK);
+    }
+    
+    /*
+      _    _      _                   __  __      _   _               _     
+	 | |  | |    | |                 |  \/  |    | | | |             | |    
+	 | |__| | ___| |_ __   ___ _ __  | \  / | ___| |_| |__   ___   __| |___ 
+	 |  __  |/ _ \ | '_ \ / _ \ '__| | |\/| |/ _ \ __| '_ \ / _ \ / _` / __|
+	 | |  | |  __/ | |_) |  __/ |    | |  | |  __/ |_| | | | (_) | (_| \__ \
+	 |_|  |_|\___|_| .__/ \___|_|    |_|  |_|\___|\__|_| |_|\___/ \__,_|___/
+	               | |                                                      
+	               |_|                             
+     */
     private Income convert(final IncomeData income) {
     	final Income newIncome = new Income();
     	newIncome.setId(income.getIncomeId());
@@ -233,12 +291,16 @@ public class IncomesApiController implements IncomesApi {
     private List<Income> convert(final List<IncomeData> incomes) {
     	final List<Income> incomeList = new ArrayList<>();
     	
-    	for (IncomeData income : incomes) {
+    	for (final IncomeData income : incomes) {
     		final Income newIncome = convert(income);
     		
     		incomeList.add(newIncome);
     	}
     	
     	return incomeList;
+    }
+    
+    private void logError(final Error error) {
+    	logger.error("[{}] {} : {}", error.getCode(), error.getMessage(), error.getDescription());
     }
 }
